feat: Implement auth for TimeEntry and Project Controllers

+++ app/Http/Controllers/ProjectController.php
@@ -2,23 +2,26 @@
 
 namespace App\Http\Controllers;
 use App\Models\Project;
-use App\Models\Client;
 
 use Illuminate\Http\Request;
 
 class ProjectController extends Controller
 {
-    //todo: auth -> index, store, show, update, destroy
-    public function index()
+    public function index(Request $request)
     {
+        $userId = $request->user()->id;
+        $projects = Project::whereHas('client', function ($query) use ($userId) {
+            $query->where('user_id', $userId);
+        })->get();
 
-        $projects = Project::all();
         return response()->json($projects, 200);
 
     }
 
     public function store(Request $request)
     {
+        $user = $request->user();
+
         $validatedData = $request->validate([
             'client_id' => ['required', 'integer', 'exists:clients,id'],
             'name' => ['required', 'string', 'max:255'],
@@ -28,20 +31,38 @@ class ProjectController extends Controller
             'end_date' => ['nullable', 'date', 'after_or_equal:start_date'],
         ]);
 
+        $clientId = $validatedData['client_id'];
+
+        $clientBelongsToUser = $user->clients()->where('id', $clientId)->exists();
+
+        if (!$clientBelongsToUser) {
+            abort(403, 'Unauthorised action. The selected client does not belong to you.');
+
+        }
         $project = Project::create($validatedData);
 
         return response()->json($project, 201);
 
     }
 
-    public function show(Project $project)
+    public function show(Project $project, Request $request)
     {
+        $userId = $request->user()->id;
+        if ($project->client->user_id !== $userId) {
+            abort(403, 'You do not own this project.');
+        }
         return response()->json($project, 200);
 
     }
 
     public function update(Project $project, Request $request)
     {
+        $userId = $request->user()->id;
+
+        if ($project->client->user_id !== $userId) {
+            abort(403, 'Unauthorised action. The selected project does not belong to you.');
+        }
+
         $validatedData = $request->validate([
             'client_id' => ['sometimes', 'integer', 'exists:clients,id'],
             'name' => ['sometimes', 'string', 'max:255'],
@@ -51,14 +72,31 @@ class ProjectController extends Controller
             'end_date' => ['sometimes', 'nullable', 'date', 'after_or_equal:start_date'],
         ]);
 
+        if (isset($validatedData['client_id'])) {
+            $clientId = $validatedData['client_id'];
+
+            $user = $request->user();
+
+            $clientBelongsToUser = $user->clients()->where('id', $clientId)->exists();
+
+            if (!$clientBelongsToUser) {
+                abort(403, 'Unauthorised action. The selected client does not belong to you.');
+            }
+        }
         $project->update($validatedData);
 
         return response()->json($project, 200);
 
     }
 
-    public function destroy(Project $project)
+    public function destroy(Project $project, Request $request)
     {
+        $userId = $request->user()->id;
+
+        if ($project->client->user_id !== $userId) {
+            abort(403, 'Unauthorised action. The selected project does not belong to you.');
+        }
+
         $project->delete();
 
         return response()->json(null, 204);
